+++ composeApp/src/commonMain/kotlin/com/taka/runejournal/feature/journal/domain/model/TimelineItem.kt
@@ -4,26 +4,26 @@ import com.taka.runejournal.core.domain.model.DrawnRune
 import kotlin.time.Instant
 
 sealed class TimelineItem {
-  abstract val id: Int
+  abstract val id: Long
   abstract val createdAt: Instant
   abstract val notes: String?
 
   data class JournalEntry(
-    override val id: Int,
+    override val id: Long,
     override val createdAt: Instant,
     override val notes: String?,
     val imageFileName: String?
   ) : TimelineItem()
 
   data class SingleRuneReading(
-    override val id: Int,
+    override val id: Long,
     override val createdAt: Instant,
     override val notes: String?,
     val rune: DrawnRune,
   ) : TimelineItem()
 
   data class PpfRuneReading(
-    override val id: Int,
+    override val id: Long,
     override val createdAt: Instant,
     override val notes: String?,
     val pastRune: DrawnRune,
